% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{backup}
\alias{backup}
\title{Hace backup de las colecciones de base de datos en un directorio especificado}
\usage{
backup(directorio)
}
\arguments{
\item{directorio}{Nombre del directorio en el que se colocaran los archivos del backup}
}
\description{
Hace backup de las colecciones de base de datos en un directorio especificado
}
