% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{calcularDistanciaEntreDosCoordenadas}
\alias{calcularDistanciaEntreDosCoordenadas}
\title{Calcula la distancia entre dos puntos geograficos}
\usage{
calcularDistanciaEntreDosCoordenadas(lng1, lat1, lng2, lat2)
}
\arguments{
\item{lng1}{longitud 1}

\item{lng2}{longitud 2}

\item{lat2}{latitud 2}

\item{lat}{latitud 1}
}
\description{
Calcula la distancia entre dos puntos geograficos
}
