% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{cargarFichero}
\alias{cargarFichero}
\title{Carga el contenido de un archivo CSV a una collecion en la base de datos.}
\usage{
cargarFichero(ruta, nombre)
}
\arguments{
\item{ruta}{La ruta donde se encuentra el fichero que se carga a la base de datos.}

\item{nombre}{El nombre en singular de la collecion donde se guarda la data contenida en el fichero.}
}
\description{
Carga el contenido de un archivo CSV a una collecion en la base de datos.
}
