% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{cargarFicheroABaseDeDatos}
\alias{cargarFicheroABaseDeDatos}
\title{Carga el contenido de un archivo CSV a una colleción en la base de datos.}
\usage{
cargarFicheroABaseDeDatos(ruta, nombre)
}
\arguments{
\item{ruta}{La ruta dónde se encuentra el fichero que se carga a la base de datos.}

\item{nombre}{El nombre en singular de la collecion donde se guarda la data contenida en el fichero.}
}
\description{
Carga el contenido de un archivo CSV a una colleción en la base de datos.
}
