% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{cargarTodo}
\alias{cargarTodo}
\title{Carga en la base de datos, la data contenida en los archivos:
1. Planton.csv: se especifican los datos referentes a uno o mas plantones.
2. Lote.csv: se especifican los datos referentes a uno o mas lotes de cultivo de cana.
3. PuntoCaptacion.csv: se especifican los datos referentes a los puntos de captacion ubicados en un lote.
4. Mediciones.csv: se especifican las mediciones realizadas durante una sesion.
4. Sesion.csv: se especifican los datos de entrada referentes a una sesion de observacion.
5. Campana.csv: se especifican los datos de entrada referentes a una campana de investigacion.}
\usage{
cargarTodo(rutaDirectorio)
}
\arguments{
\item{rutaDirectorio}{La ruta del directorio que contiene los archivos especificados anteriormente.}
}
\description{
Carga en la base de datos, la data contenida en los archivos:
1. Planton.csv: se especifican los datos referentes a uno o mas plantones.
2. Lote.csv: se especifican los datos referentes a uno o mas lotes de cultivo de cana.
3. PuntoCaptacion.csv: se especifican los datos referentes a los puntos de captacion ubicados en un lote.
4. Mediciones.csv: se especifican las mediciones realizadas durante una sesion.
4. Sesion.csv: se especifican los datos de entrada referentes a una sesion de observacion.
5. Campana.csv: se especifican los datos de entrada referentes a una campana de investigacion.
}
