% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{filtro}
\alias{filtro}
\title{Aplica filtro sobre la data especificada}
\usage{
filtro(data, filterExp)
}
\arguments{
\item{data}{Data a filtrar}

\item{filterExp}{Filtro a aplicar}
}
\description{
Aplica filtro sobre la data especificada
}
