% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{simpleCap}
\alias{simpleCap}
\title{Cambia el primer caracter de una cadena de caracteres a mayuscula.}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{la cadena que se recibira los cambios.}
}
\value{
una cadena de caracteres con el primer caracter en mayuscula.
}
\description{
Cambia el primer caracter de una cadena de caracteres a mayuscula.
}
