% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarCampana}
\alias{validarCampana}
\title{Valida los datos de una campana}
\usage{
validarCampana(fichero)
}
\arguments{
\item{fichero}{Objeto del fichero cargado}
}
\description{
Valida los datos de una campana
}
