% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarCod}
\alias{validarCod}
\title{Realiza las validaciones del id de un registro}
\usage{
validarCod(nombre, cod)
}
\arguments{
\item{nombre}{Nombre de la coleccion en la que se validara si el id existe}

\item{cod}{ID a validar}
}
\description{
Realiza las validaciones del id de un registro
}
