% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarEnLista}
\alias{validarEnLista}
\title{Valida si un valor existe en la lista de valores permitidos}
\usage{
validarEnLista(valores, punto)
}
\arguments{
\item{valores}{Valores permitidos}

\item{punto}{Valor a validar}
}
\description{
Valida si un valor existe en la lista de valores permitidos
}
