% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarFichero}
\alias{validarFichero}
\title{Hace las validaciones del fichero}
\usage{
validarFichero(fichero, tipo)
}
\arguments{
\item{fichero}{Objeto del fichero cargado}

\item{tipo}{Tipo de fichero: lote, sesion, etc}
}
\description{
Hace las validaciones del fichero
}
