% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarMedicion}
\alias{validarMedicion}
\title{Valida los datos de las mediciones}
\usage{
validarMedicion(fichero)
}
\arguments{
\item{fichero}{Objeto del fichero cargado}
}
\description{
Valida los datos de las mediciones
}
