% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarNumero}
\alias{validarNumero}
\title{Realiza las validaciones de un numero}
\usage{
validarNumero(num)
}
\arguments{
\item{num}{Valor a validar}
}
\description{
Realiza las validaciones de un numero
}
