% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarNumeroBoolean}
\alias{validarNumeroBoolean}
\title{Realiza las validaciones de un numero como booleano (0 y 1)}
\usage{
validarNumeroBoolean(num)
}
\arguments{
\item{num}{Valor a validar}
}
\description{
Realiza las validaciones de un numero como booleano (0 y 1)
}
