% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarPlanton}
\alias{validarPlanton}
\title{Valida los datos de un planton}
\usage{
validarPlanton(fichero)
}
\arguments{
\item{fichero}{Objeto del fichero cargado}
}
\description{
Valida los datos de un planton
}
