% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarPositivos}
\alias{validarPositivos}
\title{Realiza las validaciones de un numero positivo}
\usage{
validarPositivos(num)
}
\arguments{
\item{num}{Valor a validar}
}
\description{
Realiza las validaciones de un numero positivo
}
