% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarPuntoCaptacion}
\alias{validarPuntoCaptacion}
\title{Valida los datos de un punto de captacion}
\usage{
validarPuntoCaptacion(fichero)
}
\arguments{
\item{fichero}{Objeto del fichero cargado}
}
\description{
Valida los datos de un punto de captacion
}
