% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar.R
\name{validarSesion}
\alias{validarSesion}
\title{Valida los datos de una sesion}
\usage{
validarSesion(fichero)
}
\arguments{
\item{fichero}{Objeto del fichero cargado}
}
\description{
Valida los datos de una sesion
}
